audit(recovery): improve comments and move GetIBCDenomDestinationIdentifiers method
