+++ x/recovery/keeper/ibc_callbacks.go
@@ -11,6 +11,7 @@ import (
 	transfertypes "github.com/cosmos/ibc-go/v3/modules/apps/transfer/types"
 	clienttypes "github.com/cosmos/ibc-go/v3/modules/core/02-client/types"
 	channeltypes "github.com/cosmos/ibc-go/v3/modules/core/04-channel/types"
+	host "github.com/cosmos/ibc-go/v3/modules/core/24-host"
 	"github.com/cosmos/ibc-go/v3/modules/core/exported"
 
 	"github.com/tharsis/evmos/v3/ibc"
@@ -37,14 +38,14 @@ func (k Keeper) OnRecvPacket(
 	params := k.GetParams(ctx)
 	claimsParams := k.claimsKeeper.GetParams(ctx)
 
-	// check channels from this chain (i.e destination)
+	// Check channels from this chain (i.e destination).
+	// Return original ACK if:
+	// - recovery is disabled globally
+	// - channel is not authorized
+	// - channel is an EVM channel
 	if !params.EnableRecovery ||
 		!claimsParams.IsAuthorizedChannel(packet.DestinationChannel) ||
 		claimsParams.IsEVMChannel(packet.DestinationChannel) {
-		// return original ACK if:
-		// - recovery is disabled globally
-		// - channel is not authorized
-		// - channel is an EVM channel
 		return ack
 	}
 
@@ -64,9 +65,8 @@ func (k Keeper) OnRecvPacket(
 		)
 	}
 
-	// case 1: sender ≠ recipient.
-	// Recovery is only possible for addresses in which the sender = recipient
-	// (i.e transferring to your own account in Evmos).
+	// Check if sender == recipient, as recovery is only possible for transfers to
+	// a sender's own account in Evmos
 	if !sender.Equals(recipient) {
 		// Continue to the next IBC middleware by returning the original ACK.
 		return ack
@@ -76,34 +76,31 @@ func (k Keeper) OnRecvPacket(
 	account := k.accountKeeper.GetAccount(ctx, sender)
 
 	// recovery is not supported for vesting or module accounts
-	_, isVestingAcc := account.(vestexported.VestingAccount)
-	if isVestingAcc {
+	if _, isVestingAcc := account.(vestexported.VestingAccount); isVestingAcc {
 		return ack
 	}
 
-	_, isModuleAccount := account.(authtypes.ModuleAccountI)
-	if isModuleAccount {
+	if _, isModuleAccount := account.(authtypes.ModuleAccountI); isModuleAccount {
 		return ack
 	}
 
-	// Case 2. sender pubkey is a supported key (eth_secp256k1, amino multisig, ed25519)
-	// ==> Continue and return success ACK as the funds are not stuck on chain
-	if account != nil &&
-		account.GetPubKey() != nil &&
+	// Check if sender pubkey is a supported key (eth_secp256k1, amino multisig,
+	// ed25519). Continue and return success ACK as the funds are not stuck on
+	// chain for supported keys
+	if account != nil && account.GetPubKey() != nil &&
 		evmos.IsSupportedKey(account.GetPubKey()) {
 		return ack
 	}
 
-	// NOTE: Since destination channel is authorized and not from an EVM chain, we know that
-	// only secp256k1 keys are supported in the source chain. This means that we can now
-	// initiate the recovery logic
-
-	// transfer the balance back to the sender address
+	// Perform Recovery to transfer the balance back to the sender address.
+	// NOTE: Since destination channel is authorized and not from an EVM chain, we
+	// know that only secp256k1 keys are supported in the source chain. This means
+	// that we can now initiate the recovery logic
 	destPort := packet.DestinationPort
 	destChannel := packet.DestinationChannel
 	balances := sdk.Coins{}
 
-	// iterate over all the tokens owned by the address (i.e sender balance) and
+	// iterate over all tokens owned by the address (i.e sender balance) and
 	// transfer them to the original sender address in the source chain (if
 	// applicable, see cases for IBC vouchers below).
 	k.bankKeeper.IterateAccountBalances(ctx, sender, func(coin sdk.Coin) (stop bool) {
@@ -139,7 +136,7 @@ func (k Keeper) OnRecvPacket(
 		// NOTE: Don't use the consensus state because it may become unreliable if updates slow down
 		timeout := uint64(ctx.BlockTime().Add(params.PacketTimeoutDuration).UnixNano())
 
-		// Recovery the tokens to the bech32 prefixed address of the source chain
+		// Recover the tokens to the bech32 prefixed address of the source chain
 		err = k.transferKeeper.SendTransfer(
 			ctx,
 			packet.DestinationPort,    // packet destination port is now the source
@@ -212,4 +209,72 @@ func (k Keeper) OnRecvPacket(
 
 	// return original acknowledgement
 	return ack
+}
+
+// GetIBCDenomDestinationIdentifiers returns the destination port and channel of the IBC denomination,
+// i.e port and channel on Evmos for the voucher. It returns an error if:
+//  - the denomination is invalid
+//  - the denom trace is not found on the store
+//  - destination port or channel ID are invalid
+func (k Keeper) GetIBCDenomDestinationIdentifiers(ctx sdk.Context, denom, sender string) (destinationPort, destinationChannel string, err error) {
+	ibcDenom := strings.SplitN(denom, "/", 2)
+	if len(ibcDenom) < 2 {
+		return "", "", sdkerrors.Wrap(transfertypes.ErrInvalidDenomForTransfer, denom)
+	}
+
+	hash, err := transfertypes.ParseHexHash(ibcDenom[1])
+	if err != nil {
+		return "", "", sdkerrors.Wrapf(
+			err,
+			"failed to recover IBC vouchers back to sender '%s' in the corresponding IBC chain", sender,
+		)
+	}
+
+	denomTrace, found := k.transferKeeper.GetDenomTrace(ctx, hash)
+	if !found {
+		return "", "", sdkerrors.Wrapf(
+			transfertypes.ErrTraceNotFound,
+			"failed to recover IBC vouchers back to sender '%s' in the corresponding IBC chain", sender,
+		)
+	}
+
+	path := strings.Split(denomTrace.Path, "/")
+	if len(path)%2 != 0 {
+		// safety check: shouldn't occur
+		return "", "", sdkerrors.Wrapf(
+			transfertypes.ErrInvalidDenomForTransfer,
+			"invalid denom (%s) trace path %s", denomTrace.BaseDenom, denomTrace.Path,
+		)
+	}
+
+	destinationPort = path[0]
+	destinationChannel = path[1]
+
+	_, found = k.channelKeeper.GetChannel(ctx, destinationPort, destinationChannel)
+	if !found {
+		return "", "", sdkerrors.Wrapf(
+			channeltypes.ErrChannelNotFound,
+			"port ID %s, channel ID %s", destinationPort, destinationChannel,
+		)
+	}
+
+	// NOTE: optimistic handshakes could cause unforeseen issues.
+	// Safety check: verify that the destination port and channel are valid
+	if err := host.PortIdentifierValidator(destinationPort); err != nil {
+		// shouldn't occur
+		return "", "", sdkerrors.Wrapf(
+			host.ErrInvalidID,
+			"invalid port ID '%s': %s", destinationPort, err.Error(),
+		)
+	}
+
+	if err := host.ChannelIdentifierValidator(destinationChannel); err != nil {
+		// shouldn't occur
+		return "", "", sdkerrors.Wrapf(
+			channeltypes.ErrInvalidChannelIdentifier,
+			"channel ID '%s': %s", destinationChannel, err.Error(),
+		)
+	}
+
+	return destinationPort, destinationChannel, nil
 }
